requirements.txt
flask
openai
python-dotenv
google-generativeai
